seperate controller logic and service logic
