+++ src/app/api/papers/route.ts
@@ -1,42 +1,29 @@
 import { NextResponse, type NextRequest } from "next/server";
-import { connectToDatabase } from "@/lib/database/mongoose";
-import Paper from "@/db/papers";
-import { type IPaper } from "@/interface";
-import { escapeRegExp } from "@/lib/utils/regex";
-import { extractUniqueValues } from "@/lib/utils/paper-aggregation";
+import { getPapersBySubject } from "@/lib/services/paper";
 
 export const dynamic = "force-dynamic";
 
 export async function GET(req: NextRequest) {
   try {
-    await connectToDatabase();
+
     const url = req.nextUrl.searchParams;
-    const subject = url.get("subject");
+    const sub = url.get("subject");
 
-    if (!subject) {
-      return NextResponse.json(
-        { message: "Subject query parameter is required" },
-        { status: 400 },
-      );
-    }
-
-    const escapedSubject = escapeRegExp(subject);
-    const papers: IPaper[] = await Paper.find({
-      subject: { $regex: new RegExp(`${escapedSubject}`, "i") },
-    });
-
-    const uniqueValues = extractUniqueValues(papers);
+    const paper = await getPapersBySubject(sub as string);
 
     return NextResponse.json(
-      {
-        papers,
-        ...uniqueValues,
-      },
+      paper,
       { status: 200 },
     );
   } catch (error) {
+    if (error instanceof Error) {
+      return NextResponse.json(
+        { message: "Failed to fetch papers", error: error.message },
+        { status: error.message === "Subject query parameter is required" ? 400 : 500 },
+      );
+    }
     return NextResponse.json(
-      { message: "Failed to fetch papers", error },
+      { message: "Failed to fetch papers", error},
       { status: 500 },
     );
   }
